Force netty-tcnative-boringssl-static to 2.0.45.Final
When upgrading the lib imapnio-core to 4.3.7 on gatling-imap submodule, it creates an issue in james-gatling.

It can't find SSLPrivateKeyMethod, because Gatling is using it through an older version of Netty than imapnio. Forcing an upgrade on it seems to solve it for now.

But we should upgrade our Gatling lib at some point...