Remove default headers in Http protocol definition
After the upgrade from Gatling 3.0.3 to 3.4.2, it seems those headers are always being injected on top (maybe before they were overrided), thus James was fetching the first one (default) and not the one overrided in the request.

Thus that creates issues with jmap-rfc simulations that rely on the Accept header to use jmap-rfc and not draft.

After scanning the code and usage, it seems safe to me to just remove those.